Extract test helper functions from ANNSuite into separate object
